input-feeder/requirements.txt
paho-mqtt
asyncio
aiofiles
pyyaml